Multi Segment Progress Bar with modular stripes
