+++ example/pubspec.yaml
@@ -1,23 +1,17 @@
-name: multi_segment_linear_indicator
-description: Example app for the multi_segment_linear_indicator package.
-version: 1.0.0
-publish_to: none
+name: percent_indicator_example
+description: A new Flutter project.
 
-environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.0.0"
+version: 1.0.0+1
+publish_to: "none"
 
 dependencies:
   flutter:
     sdk: flutter
-  multi_segment_linear_indicator:
+
+  percent_indicator:
     path: ../
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^3.0.1
-
+environment:
+  sdk: ">=2.12.0-0 <3.0.0"
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
